+++ idl/favorite.thrift
@@ -3,8 +3,8 @@ namespace go favorite
 include "video.thrift"
 
 struct FavoriteActionRequest {
-    1: i32 user_id, // 用户id
-    2: i32 video_id, // 视频id
+    1: i64 user_id, // 用户id
+    2: i64 video_id, // 视频id
     3: i32 action_type, // 1-点赞，2-取消点赞
 }
 
@@ -14,8 +14,8 @@ struct FavoriteActionResponse {
 }
 
 struct FavoriteListRequest {
-    1: i32 action_id, // 当前操作用户的用户id
-    2: i32 user_id,   //列出user_id点赞的视频
+    1: i64 action_id, // 当前操作用户的用户id
+    2: i64 user_id,   //列出user_id点赞的视频
 }
 
 struct FavoriteListResponse {
@@ -25,16 +25,16 @@ struct FavoriteListResponse {
 }
 
 struct IsUserFavoriteRequest{
-    1: i32 user_id,  // 当前操作用户的用户id
-    2: i32 video_id, // 视频id
+    1: i64 user_id,  // 当前操作用户的用户id
+    2: i64 video_id, // 视频id
 }
 
 
 service FavoriteService {
     FavoriteActionResponse FavoriteAction(1: FavoriteActionRequest Request),
     FavoriteListResponse GetFavoriteList(1: FavoriteListRequest Request),
-    i32 GetVideoFavoriteCount(1: i32 video_id),//获取video_id的点赞总数
-    i32 GetUserFavoriteCount(1:i32 user_id), //获取user_id的点赞数
-    i32 GetUserTotalFavoritedCount(1:i32 user_id), //获取user_id的总获赞数量
+    i32 GetVideoFavoriteCount(1: i64 video_id),//获取video_id的点赞总数
+    i32 GetUserFavoriteCount(1:i64 user_id), //获取user_id的点赞数
+    i32 GetUserTotalFavoritedCount(1:i64 user_id), //获取user_id的总获赞数量
     bool IsUserFavorite(1:IsUserFavoriteRequest Request),
 }
