+++ idl/message.thrift
@@ -1,16 +1,16 @@
 namespace go message
 
 struct Message {
-    1: i32 id, // 消息id
-    2: i32 to_user_id, // 该消息接收者的id
-    3: i32 from_user_id, // 该消息发送者的id
+    1: i64 id, // 消息id
+    2: i64 to_user_id, // 该消息接收者的id
+    3: i64 from_user_id, // 该消息发送者的id
     4: string content, // 消息内容
     5: optional i64 create_time, // 消息创建时间
 }
 
 struct MessageChatRequest {
-    1: i32 from_user_id, // 发送方id
-    2: i32 to_user_id, // 对方用户id
+    1: i64 from_user_id, // 发送方id
+    2: i64 to_user_id, // 对方用户id
     3: i64 pre_msg_time, // 上次最新消息的时间（新增字段-apk更新中）
 }
 
@@ -21,8 +21,8 @@ struct MessageChatResponse {
 }
 
 struct MessageActionRequest {
-    1: i32 from_user_id, // 用户id
-    2: i32 to_user_id, // 对方用户id
+    1: i64 from_user_id, // 用户id
+    2: i64 to_user_id, // 对方用户id
     3: i32 action_type, // 1-发送消息
     4: string content, // 消息内容
 }
